Separated mms and psApi handlers
